Implement basic action and trigger factories (#5)
