+++ tests/unit/StateMachine/ConversationStateTest.php
@@ -3,7 +3,7 @@ declare(strict_types=1);
 
 namespace FireGento\MageBot\StateMachine;
 
-use FireGento\MageBot\StateMachine\Serialization\NewActionFactory;
+use FireGento\MageBot\StateMachine\Serialization\FakeActionFactory;
 use FireGento\MageBot\StateMachine\Serialization\SerializableAction;
 use PHPUnit\Framework\TestCase;
 
@@ -11,7 +11,7 @@ class ConversationStateTest extends TestCase
 {
     protected function setUp()
     {
-        SerializableAction::setActionFactory(new NewActionFactory());
+        SerializableAction::setActionFactory(new FakeActionFactory());
     }
 
     protected function tearDown()
